Move parallel commands into submodule
